rf_radiated: call EnterMFGMode() in runTest() instead of setUp()
Vendors' MFG software are often unreliable, some times we have to retry calling
EnterMFGMode(). It is thus more reasonable to put it in runTest() instead of
setUp() because failures in setUp() should be considered critical, and normally
cannot be recovered.

Also fixed a bug that always raise exception in setUp().

BUG=None
TEST=Manually tested

Change-Id: I1defe9e84bd35f9660a24bb313453699e0596178
Reviewed-on: https://chromium-review.googlesource.com/304971
Commit-Ready: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>