bluetooth: support enable_magnet arg
The magnet enabling task can now be an individual task by specifying
the enable_magnet arg. It is advantageous to make it a standalone task
instead of attaching it to other tests. In this way, it is obvious to
see when the magnet is enabled. It is also useful for debugging purpose
if one wants to execute this task in the engineering mode.

BUG=chrome-os-partner:47419
TEST=manually tested.

Change-Id: I7a21b9d4a518656a7ea435507a81806b42f13bea
Reviewed-on: https://chromium-review.googlesource.com/311944
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Tony Lin <tonycwlin@chromium.org>
Reviewed-by: Pin-chih Lin <johnylin@chromium.org>