bluetooth: do not control magnet in middle
To adapt to the characteristics of the new base firmware, resetting
a base may need up to 10 seconds to get stabilized. Besides, the
wake up time of a base has been lengthened such that there is no
need to reset a base too frequently. Hence, this patch removes
some of the electro-magnet enabling/disabling actions.

BUG=chrome-os-partner:47419
TEST=manually tested.

Change-Id: I14f78738efe0ba901d2104a85750e0807975e057
Reviewed-on: https://chromium-review.googlesource.com/311945
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Tony Lin <tonycwlin@chromium.org>
Reviewed-by: Pin-chih Lin <johnylin@chromium.org>