overlord: refactor UI and extend dashboard functionality.
Show "UI" button in ClientBox if the fixture has "ui" in its "context"
properties, on click, it will open a FixtureWindow, which is also used
in Fixture dashboard.
Also, due to issue 429419, we can't open more than 30 secure websockets
in the same time, we automatically close the oldest fixture window if
there are too many fixture windows (we compute the number of maximum
fixture windows according to the window size).

BUG=None
TEST=Manual test

Change-Id: I0886b81263a7cff331f546f5404bad0bd0a78398
Reviewed-on: https://chromium-review.googlesource.com/307843
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>