bluetooth: CheckDisconnectionOfPairedDeviceTask
This patch adds a new task of checking if a paired device has
disconnected from the test host.

When a user pressed SHIFT-P-A-I-R after completing the keyboard
test, the base would disconnect from the test host. This task
is used to check if the user has actually pressed the key combo.

BUG=chrome-os-partner:47419
TEST=manually tested.

Change-Id: I3b0fc2db2ee4f34848cb960f227d64a26cad2cfc
Reviewed-on: https://chromium-review.googlesource.com/311953
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Pin-chih Lin <johnylin@chromium.org>