overlord: TerminalWindow: show disconnection overlay when connection dropped
Although terminal window should disappear when client exit, there are
some case that the connection drop unexpectedly. In such case we show a
'connection lost' overlay on top of the terminal window. An user can
click at the overlay to dismiss it and look at the terminal content
before the connection was lost.

BUG=none
TEST=manually

Change-Id: Ie3ad9a6d1c7a746b491f8ef4ad57740f3489e1af
Reviewed-on: https://chromium-review.googlesource.com/309691
Commit-Ready: Wei-Ning Huang <wnhuang@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>