gps: wait for glgps to start writing to nmea_out pipe
There were some instances where running GPS test was not able
to get any output from the nmea_out pipe.  It turns out that
the order of (1) glgps starts writing to the pipe, and (2)
GPS test starts reading from the pipe is significant.  So we
add a command to check that output is already being written
to nmea_out.

BUG=None
TEST=Manually on DUT

Change-Id: Ib72b2046ab29ab144b40a07c071524a3cbc17f26
Reviewed-on: https://chromium-review.googlesource.com/308537
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>