overlord: replace forceUpdate by setState
According to React's document, "this.state" object should be treated as
an immuable object. So the previous implementation that change some part
of this.state and then "this.forceUpdate()" would be error-prone.
For example:

removeAllClients: function () {
  this.setState({clients: []});
},
addClient: function (client) {
  this.state.clients.push(client);
  this.forceUpdate();
}

There might be a case that right after "removeAllClients" is called,
"addClient" is also called. But since setState is an asynchronous
function, this.state.client might not be an empty array, might still in
previous state, and cause wrong behavior.

The correct way:
  this.state.value = new_value; // new_value doesn't depend on current value
    =>  this.setState({value: new_value});
  this.state.value += 1; // new value depends on current value
    =>  this.setState(function (state, props) {
          // state and props are the current this.state and this.props
          // right before this anonymous function is called repectively.
          return {value: state.value + 1};
        });
  this.state.clients.push(new_client); this.forceUpdate();
    => this.setState(function (state, props) {
          state.clients.push(new_client);
          // no return value since we are not assigning new values to
          // any attribute.
       });

BUG=chromium:551264
TEST=Manual test

Change-Id: I4812b056b911a336f9cf821e7b88a9a3d50e71ec
Reviewed-on: https://chromium-review.googlesource.com/310690
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>