move cros.factory.rf to cros.factory.test.rf.
The 'rf' folder was for radio-frequency related testing fixtures and we should
move them to cros.factory.test.rf.

It's not 'system' or 'board' because its code was not for DUT to run (it's for
the external fixtures). Note some modules, for example the 'modem', may need to
run on DUT and we will move that in future.

BUG=chromium:557573
TEST=make test

Change-Id: I13780f0df596286e6e3370207eec3c851e1362c5
Reviewed-on: https://chromium-review.googlesource.com/313933
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>