Move SystemInfo and SystemStatus from cros.factory.system to sub module 'state'.
The SystemInfo and SystemStatus has too many dependencies that would make
'system' not able to contain most base classes and would be a problem when we
want to make all system modules to share same base class.

To reduce extra dependency, move the two functions to a new module 'state'.

The 'SetBacklightBrightness' is left unchanged and will be revised when we have
finished the transition to DUT-aware system modules.

BUG=chromium:557573
TEST=make test

Change-Id: I54993fac25fa4585ac83ded28a99278ccf4a1ba7
Reviewed-on: https://chromium-review.googlesource.com/313904
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>