factory: make DUT an argument of system.GetBoard
In fixture-based testing, a Board instance needs to know how to connect
to the DUT. So we make DUT an argument to system.GetBoard function, and
system.GetBoard will pass the DUT object to the Board instance. This
argument is optional, when DUT is not given, we will use LocalTarget.

CQ-DEPEND=CL:313935
BUG=chromium:551806
TEST=manual test

Change-Id: I8a5f1f1b5021fef770087eedc26fdd3c4c6d60c0
Reviewed-on: https://chromium-review.googlesource.com/316760
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>