[Battery Current] Support testing multiple USB Type C ports
We use battery_current.py to test USB TypeC charge function. Because the
test time is short. So we add USBTypeC status check in the test case, it
can check if we connect it by a power source and specify the minimum
voltage from the power.

BUG=None
TEST=manual test

Change-Id: Ia8deae0c79de3a9818fe33eda53a5921db6f970b
Reviewed-on: https://chromium-review.googlesource.com/318092
Commit-Ready: Hsu Wei-Cheng <mojahsu@chromium.org>
Tested-by: Hsu Wei-Cheng <mojahsu@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>