dut: Fix 'dut.temp' module on ChromeOS.
The 'dut.temp' module was based on mktemp(1), which was not clearly defined in
UNIX utility spec. BSD, Linux, and Android all have their own syntax that is not
fully compatible.

The solution is to create an AndroidTemporaryFiles implementation for dut.board
to select right modules.

BUG=none
TEST=make test

Change-Id: I4f061107a3a7ef8cc805f00724f4578d5621561a
Reviewed-on: https://chromium-review.googlesource.com/319301
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>