Move test managers to cros.factory.test.utils.
We are deprecating "system" folder because any components should be either a
component driver (DUT-aware API) or a software stack based on components.
And for the later, they should live in cros.factory.test. The utils folder
should be the best place for them.

BUG=chromium:557573
TEST=make test

Change-Id: I8c1b7d402b8b22f14d49febde215543cbcff900d
Reviewed-on: https://chromium-review.googlesource.com/319692
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Shun-Hsing Ou <shunhsingou@chromium.org>