tools: Move system/disk_space to tools/disk_space.
The 'disk_space' module is currently used only by Goofy to obtain the disk usage
on system (primarily for preserving log space), and not really needed to remote
DUT execution.

To simplify migration, we should put it in 'tools' so only Goofy runts is for
local host.

BUG=chromium:557573
TEST=make test

Change-Id: I8bd1e1656e1d934d412d78acac682b3eb1db65fb
Reviewed-on: https://chromium-review.googlesource.com/319727
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>