dut: Add DUT-aware "bluetooth" component.
Move the bluetooth module from system to DUT folder. The bluetooth module is
currently using DBus so it can't run remotely.

The API was also changed to remove the argument (host_mac) in constructor
because it's only used in GetFirstAdapter and GetAdapters, which can be made as
the argument for those two functions (and is more flexible if a test wants to
use multiple bluetooth host adapters).

BUG=chromium:557573
TEST=make test

Change-Id: I967fa5a18e534ffb9a9bbac850544eeecde490c4
Reviewed-on: https://chromium-review.googlesource.com/320068
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>