Fix race condition for event socket read
We have 2 threads which read event, one is EventServerRecvThread
the other is from test_ui.py with the EVENT_LOOP_WAIT event type.
The test will be hang sometimes if the goofy:end_test event is read by
EventServerRecvThread.
We fixed it by only creating the EventServerRecvThread when the event type
is not EVENT_LOOP_WAIT.

BUG=None
TEST=run the hang test before the patch 5 times

Change-Id: I200e7770f38ea7c4e58c5b0ae505a7525c67d434
Reviewed-on: https://chromium-review.googlesource.com/320142
Commit-Ready: Shun-Hsing Ou <shunhsingou@chromium.org>
Tested-by: Shun-Hsing Ou <shunhsingou@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>