Add udev component and migrate removable_storage test to dut API
Previously removable_storage is written with pyudev, which cannot be applied in
station-based mode. To reuse the same test flow with different kinds of dut, an
udev component is added into the board to help us detecting udev events.

removable_storage test is migrated to the dut API.

In this CL only LocalUdevMonitor is implemented, which still using
pyudev to monitor the device. Implementation using other methods will be
added in another CL later.

BUG=chromium:557573
TEST=Perform SD card/USB read/write test.

Change-Id: Ib6aa2d79738f499362b796e078ef278b7e2c0eb7
Reviewed-on: https://chromium-review.googlesource.com/320183
Commit-Ready: Shun-Hsing Ou <shunhsingou@chromium.org>
Tested-by: Shun-Hsing Ou <shunhsingou@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>