dut: Fix temp module on Android.
When dir is not specified, we want to use the default temp folder. On Android
this implies we need TMPDIR to be set. Unfortunately when using ADB link (adb
shell) TMPDIR will be set only in interactive shells. This will cause mktemp to
fail if we run it as "adb shell mktemp".

To solve this, we need to set explicitly a default path for all Android boards.
The new TMPDIR variable will be used when calling mktemp with dir as None.

BUG=chromium:557573
TEST=unittest and verified manually:
 from cros.factory.test import dut
 d = dut.Create(board_class='AndroidBoard', link_class='ADBLink')
 print d.temp.mktemp(False)

Change-Id: I583275ecb2d016c881413450ca2ae6da012fe909
Reviewed-on: https://chromium-review.googlesource.com/320035
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>