dut: Add SendFile API to copy files into DUT.
It seems easier to understand if we have a SendFile wrapper in
comparison to dut.link.Push, when there are lots of ReadFile and WriteFile
calls in same code segment.

Additionally, the SendFile may be extended to support syncing multiple files
efficiently (for example doing rsync if the board supports it), or allow
setting additional attributes.

BUG=chromium:557573
TEST=make test

Change-Id: I4c7cd262086473472f1df368925a4df183aed775
Reviewed-on: https://chromium-review.googlesource.com/320547
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>