dut: Prevent unexpected stderr messages in ADB stdout.
The ADB protocol seem to have mixed stderr and stdout in same output channel so
we can't simply grab stderr or ADB command (which will only get trace logs, if
requested).

Currently there is no good solution yet so a quick solution is to always
redirect stderr to /dev/null if that is not needed. In future we may improve
that by utilizing temporary files.

BUG=chromium:557573
TEST=make test

Change-Id: I0bbce8519c604212c178cdcf0c13334edf773db6
Reviewed-on: https://chromium-review.googlesource.com/320452
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Shun-Hsing Ou <shunhsingou@chromium.org>
Reviewed-by: Shun-Hsing Ou <shunhsingou@chromium.org>