system: Move MSR to thermal_slope test.
The MSR (Model Specific Register) module is currently only used to retrieve
energy status for thermal_slope test.  It is possible to provide such
information in dut.status module but (1) It's very specific to x86 and no
different per-board implementations (2) We don't need anything else from it yet,
so a simpler solution is to just move MSRSnapshot into thermal slope test
itself. If we need to use MSR in future we can move it to dut.status with right
usage.

BUG=chromium:557573
TEST=make test

Change-Id: I42ae09cba706e470d06471f6cc309bdcceae688c
Reviewed-on: https://chromium-review.googlesource.com/320525
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>