hwid_cmdline: add json_output argument.
In order to get the output of the hwid methods through stdout of the shell
command, we add an argument 'json_output' for this purpose. When this arguement
is enabled, the command should only output the returned value of the method to
stdout by json format.

BUG=chromium:557573
TEST=make test

Change-Id: I4957a4b0c0c62514e17db7059ba3403dd22cd16d
Reviewed-on: https://chromium-review.googlesource.com/321591
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>