process_utils: Add RedirectStandardStreams.
For those methods that use the shell command line as API, we should only print
the returned value of the methods and avoid to output other message to stdout.
We implement the RedirectStandardStreams that can let us redirect standard
streams to other file object.

BUG=chromium:557573
TEST=make test

Change-Id: I81b0df1ed9d47d27c56de1a09c79ed2aa7a04d61
Reviewed-on: https://chromium-review.googlesource.com/320918
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>