pytest: Remove the binaries in touch calibration boards folder
To reduce factory archive size, we want to only include the binaries
when needed. The utilities in touchscreen calibration should live in its
own board overlay (or CPFE).

Existing tools are already in Ryu or Samus overlay so we're removing the
files from ToT.

BUG=none
TEST=make test

Change-Id: Icdeaa4b58e96d4f7a73b125d088a21bf81d2262f
Reviewed-on: https://chromium-review.googlesource.com/322083
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>