fmap: select which fmap to decode
Decode the correct fmap by checking the fmap name in order to
distinguish between the EC fmap and the main fmap.

Multiple flashmaps can be included in a single image. It can no longer
be trusted that the main fmap is the first fmap region in the image, so
fmap_decode should check the fmap name before continuing.

BUG=chromium:470407
BRANCH=none
TEST=Add ec.RO.bin to the bios in a region before ro-fmap then
'emerge-veyron_jerry chromeos-bootimage depthcharge'

Change-Id: I13738e570d4ad25a588609931e5492880f8387cf
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/322262
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>