overlord: TerminalWindow: add upload error indicator UI
Sending a GET against the upload path checks if the upload could be
potentially erroneous without actually sending the POST request and wait
for it to finish. The overlord CLI already make use of this API.
Implement this in the web UI so the user can be notified of error.

BUG=chromium:465674
TEST=manually

Change-Id: Ie69bf35c6a278c6a5174907f6aafb1e766026893
Reviewed-on: https://chromium-review.googlesource.com/322521
Commit-Ready: Wei-Ning Huang <wnhuang@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>