overlord: ghost: buffer handling of leftover data
When the overlord control channel switch into a data channel after
registration, the leftover data need to be interpreted as input data
steam. Instead of having to deal with the leftover separately, implement
a wrapper class that supports 'UnRecv'. UnRecv push the data in to a
buffer, which will later be consume with Recv() is called.

BUG=none
TEST=`go/src/overlord/test/overlord_e2e_unittest.py`

Change-Id: Ia3068528ceeded5c49f8ea905fb7cc8cedb22a1f
Reviewed-on: https://chromium-review.googlesource.com/323923
Commit-Ready: Wei-Ning Huang <wnhuang@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Reviewed-by: Rong Chang <rongchang@chromium.org>