overlord: revert BufferedConn implementation for the go version
The UnRead method is introduce in commit a28cd2, which turn out to be a
bad idea for the go version. Since go uses a separate go routine for
reading data. The UnRead method will cause the buffer state to be
out-of-order. If we add a mutex to it, the UnRead method will be block
all the time since the reader goroutine is always reading data.
Thus we remove the BufferedConn implementation.

BUG=none
TEST=`go/src/overlord/test/overlord_e2e_unittest.py`

Change-Id: Ie55e5cf53015d892a0c8c6412f3b2a6bb0f1acbf
Reviewed-on: https://chromium-review.googlesource.com/327130
Commit-Ready: Wei-Ning Huang <wnhuang@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Rong Chang <rongchang@chromium.org>