stressapptest: Add flag for use free memory only or total memory
For the case of tests using stressapptest for "memory validation", it
should test give ratio according to system total memory (and abort if
the condition can't meet). However for devices on run-in rack, they may
be running multiple stress tests in parallel - for example stressapptest
(for CPU & RAM), glbench (for GPU), and other tests for network or
audio. And we can't expect how much memory will be free at that time -
so using the ration for "free memory" would be the better idea.

This change adds a new argument to specify should we use only free
memory or not (i.e., total memory).

BUG=None
TEST=Run stressapptest on Lucid.

Change-Id: I3664fb669c81bcc1370c41ccc42494e8c2f4f19a
Reviewed-on: https://chromium-review.googlesource.com/327449
Commit-Ready: Shun-Hsing Ou <shunhsingou@chromium.org>
Tested-by: Shun-Hsing Ou <shunhsingou@chromium.org>
Reviewed-by: Shun-Hsing Ou <shunhsingou@chromium.org>