goofy: allow ignore file to disable test lists
Allow for a file called IGNORE in the test_lists directory.  If any
of the test list module names (filename without ".py" extension)
are contained within this list, it will be ignored by Goofy.

Useful if a test list currently contains errors, but will be fixed
and used in the future.

TEST=Manually on DUT
BUG=None
CQ-DEPEND=CL:328900

Change-Id: I3f89eac4d4424a679d8ef525c06f6329bdbee6d3
Reviewed-on: https://chromium-review.googlesource.com/328893
Commit-Ready: Joel Kitching <kitching@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>