fixture test list: scan serial number support
Station-based fixtures need to be able to keep track of the serial
number of the current device being tested, which should then be
stored into DeviceData on the Goofy device.

This adds support in the fixture test list to clear DeviceData
when the fixture_entry test is run, and also (by default) turns
the fixture-based test list into an AutomatedSequence, so the
serial numbers are always properly cleared before and after the
test list's sequence of tests are run.

Also add support to "press spacebar" before running the sequence
of tests (useful for cases when the operator still has to perform
some task before starting the station, e.g. closing an RF chamber).

BUG=chromium:551806
TEST=Manually on DUT

Change-Id: Ia14472809414cecc04a0deea6f7bdc888845b13b
Reviewed-on: https://chromium-review.googlesource.com/329485
Commit-Ready: Joel Kitching <kitching@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>