Fix import error "AttributeError: 'module' object has no attribute 'factory'"
It is found that on some environments (i.e., a fresh checkout of public
repo and totally build from source, and the first time run
build_packages), programs importing other modules in the form
"import a.b as c" would lead to AttributeError when making factory bundle.

The reason is "import a.b as c" will need to first inject 'b' into
existing name space a, which will become a problem if a is not
initialized yet (for instance, if the one imports a.b was actually
pulled in by a's own dependency or if the modules in sys.modules was
explicitly modified).

The correct form is to do "from a import b as c", which only resolves
sys.modules and no need to worry about the name space problem.

Reference: http://stackoverflow.com/questions/24807434/ .

BUG=none
TEST=make test; make bundle

Change-Id: I8550e55ef98c5322228f3229ad3012cafcf4cc96
Reviewed-on: https://chromium-review.googlesource.com/329833
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>