wifi_throughput: update to use new wifi system module
Update wifi_throughput test to use DUT API WiFi system module.  Some
logic (such as ensuring that the network is disconnected) can be removed
from the test, since it is dealt with in the WiFi module.

Add parameters to control (1) whether LED lights are set to blink, and
(2) which WLAN interface to use for the test.

Update some strings to be Unicode-friendly, for when test configuration
contains a SSID with UTF-8 characters.

Remove ambiguous "signal strength" and rename to "strength", matching the
corresponding value in WiFi system module.

BUG=None
TEST=Locally with Chromebook

Change-Id: I45fd48ffec3ae0be74655eb9a5bfe4c5c1ec8924
Reviewed-on: https://chromium-review.googlesource.com/331887
Commit-Ready: Joel Kitching <kitching@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Mao Huang <littlecvr@chromium.org>