station-based: use 'with' to properly handle exceptions
In previous implementation, in order to allow for optional
AutomatedSequence, we use:
  if automated_sequence:
    auto_group.__enter__()
  # do something
  if automated_sequence:
    auto_group.__exit__(None, None, None)
to enter and exit an optional context.
However, if an exception is raised in '# do something',
auto_group.__exit__ won't be executed, and the state of
test_lists.builder_state will be wrong. The original exception will be
consumed by TestList and TestList will raise another exception. It will
be confusing and hard to debug.

In this CL, we use 'with' statement so that any exceptions will be
handled properly (e.g. they will be logged). To keep test list hierarchy
from changing, if automated_sequence is false, AutomatedSequence will be
replaced by TestGroup.

BUG=None
TEST=manual test (run goofy on chromebook)

Change-Id: I1fac856e87ae2066032fcc140997296ba44cb8fd
Reviewed-on: https://chromium-review.googlesource.com/333254
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>