offline test: fetch test result from DUT
This pytest fetch result of offline test from DUT.  It will pull file
'logfile' and 'task_id' from DUT, and check if all tasks are completed.
The log file can be uploaded to shopfloor by setting argument
'upload_to_shopfloor' to True (it's also the default value).
If log file is not uploaded to shopfloor or some offline test tasks are
not completed, content of the log file will be logged in
/var/factory/log/factory.log.

BUG=chromium:573933
TEST=make test, manual test

Change-Id: Ib01eda77c1578325285d812638cef5ec33c97fc3
Reviewed-on: https://chromium-review.googlesource.com/333814
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>