gooftool: add class TmpChroot.
For in-place wiping, we need to create a root environment in tmpfs.
The class TmpChroot will create a root envrionment under given
directory, which includes python, busybox and other given dependencies.
The TmpChroot object can be used with 'with' statement:

  with TmpChroot(...).Chroot():
    # python is under new root
  # new root is unmounted and python is under real root

Or:

  with TmpChroot(...).PivotRoot(old_root):
    # python is under new root
  # new root is unmounted and python is under real root

For pivot root, the real root is swap with new root using
pivot_root(8). An 'old_root' argument can be used to specify where to
put the real root under new root.

BUG=None
TEST=manual tested with both Chroot, PivotRoot,
     the system will recover when leaving 'with' statment

Change-Id: Ib6466c11bd79c0c4b677f1bc2373e098146c15c1
Reviewed-on: https://chromium-review.googlesource.com/336371
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Shun-Hsing Ou <shunhsingou@chromium.org>