gooftool: Fix broken Verify, Finalize command.
The positional arguments for gooftool sub commands may cause chained
commands to fail. For example, in the command chain

 finalize => verify => verify_keys,

If a parameter is defined only in verify_keys, when we invoke finalize
command then the arg in options would be missing (because the real one
gets parsed is "finalize").

The only solution is to not use positional arguments. Changed the
positional argument for verify_rootfs, verify_system_time to
'release_rootfs', and the argument for firmware in verify_keys to be
'firmware_path'.

New format:
 gooftool verify_rootfs --release_rootfs /dev/mmcblk0p3
 gooftool verify_systsem_time --release_rootfs /dev/sda3
 gooftool verify_keys --release_rootfs /dev/sda3 --firmware_path image.bin

The shell script shortcuts are also removed to prevent confusion.

BUG=chromium:573933
TEST=make test; gooftool verify_keys; gooftool verify

Change-Id: I33af20faf1ab8ba7b1fa70e9aabaad0a5c393bc7
Reviewed-on: https://chromium-review.googlesource.com/339160
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>