gooftool: Support black-listed mwifiex and refactor network device probing.
For wireless disabled by shill and not listed as 802.11 peripherals
under iwconfig (for instance, mwifiex peripherals), we have to probe by
using 'iw dev'.

The probing code is now extended to many variants that are not flimflam
(shill), so we do want to change the _FlimflamDevices to a more generic
form - the new _NetworkDevices class, with underlying implementations to
support flimflam, iwconfig, and iw.

BUG=chromium:604278
TEST=gooftool probe

Change-Id: I64f8ce8202501b7e8cd2c741c2f27f0907e4282b
Reviewed-on: https://chromium-review.googlesource.com/339533
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>