dhcp_utils: keep iface IP if possible
The IPs of interfaces managed by DHCPManager would change when factory
restart. Since the assigned IP address is considered being used.
Change the implementation of DHCPManager so that when an IP is assigned
to an interface, DHCPManager will use that IP.
However, if the assigned IP is not the first IP in the subnet
(e.g. 192.168.1.2/24), then we will reassign a new IP address.

TEST=manual test
BUG=None

Change-Id: I111a297befa14cb815be2d16d125e2203dc839aa
Reviewed-on: https://chromium-review.googlesource.com/339939
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>