network: fix GetUnmanagedInterface
Sometimes, an interface is marked with DHCP by shill, but the IP is not
set because there is no DHCP server reachable from this interface.
So when we find an interface is managed by shill (with DHCP mark), we
add an additional checking, checks if the interface had got any IP yet.
If it's not, we still assume this interface is unmanaged.

TEST=manual test
BUG=None

Change-Id: Id3815c70ae7fe245b8fcc80518cbf3b375c07e41
Reviewed-on: https://chromium-review.googlesource.com/340941
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>