net_utils: change algorithm to find unused subnet
The previous algorithm will decrease prefix bits when there is an
interface in a subnet with fewer prefix bits, and the algorithm to
avoid using subnets in "exclude_ip_prefix" is wrong.
In this CL, we reimplement the algorithm so that "preferred_prefix_bits"
is the smallest prefix bits will be used, the returned subnet can only
be smaller.

BUG=None
TEST=manual test

Change-Id: If1b36a0934889aaf4cee614e668a0d9d7d812987
Reviewed-on: https://chromium-review.googlesource.com/341292
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>