station-based: add prestart_steps for station_based_test_list
Some operations (e.g. scan serial number) better to be done before the
DUT is connected to the station. This change allows this to be possible.

We also stop clearing device data on StartStationTest, since we've
cleared it on EndStationTest. And we might want to setup device data in
'prestart_steps'.

BUG=chromium:506788
TEST=None

Change-Id: Ibd25d6821edecc87c886af3d55152eaf41593fe4
Reviewed-on: https://chromium-review.googlesource.com/341516
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>