raiden_cc_flip_check: add instructions and countdown
When we are using BFT and double CC cable to to a automatic cc flip,
we didn't show any instructions, and operators might get confused, don't
know what is going on.

We add a countdown timer if the timeout is set to non-zero and ask
operator to press space on test finished.

BUG=None
TEST=manual test

Change-Id: I276bab724aaf28d5b41230953f64c4ecd5e04c4a
Reviewed-on: https://chromium-review.googlesource.com/342805
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Pin-chih Lin <johnylin@chromium.org>