utils: Allow platform utils to load on systems without ctypes.
On some systems building ctypes may be not easy (for instance, Android)
and we do want to allow loading it only when available.

The only user of ctypes is monotonic time so a solution here is to load
ctypes only when that is invoked.

Also implemented reusing last found function to speed up timer
functions.

BUG=none
TEST=make test

Change-Id: I6953d7cdf670d04c9cdd44eabcceafd4ce2d0765
Reviewed-on: https://chromium-review.googlesource.com/343854
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Shun-Hsing Ou <shunhsingou@chromium.org>