Device-Aware API: Change from DUT API to a more general Device API.
The DUT-Aware API works fine for pytests, but we now realized there are
more to change for station-based environments - especially in the test
controller "Goofy" and test frameworks (for example test list) that
needs to deal with different environments, like offline mode or Windows.

The solution is to extend DUT API to a more general namespace that both
DUT and station can use. It also needs to live outside 'test' sub folder
because many Goofy and core modules will need it, for finding
environment settings like location of factory related folders.

Before:
  from cros.factory.test import dut
  self._dut = dut.Create()

After:
  from cros.factory.device import device_utils
  self._dut = device_utils.CreateDUTInterface()

And the new approach to specify DUT board class is now to use config:
 # chromeos-factory-board/files/config/devices.json
 { "dut": {"board_class": "DaisyBoard", "link_class": "ADBLink"} }

Note this change is backward-compatible and the migrate_dut_api script will
move board implementations to right place.

BUT=chromium:557573
TEST=make test

Change-Id: I4bf2bb22150b3a153c92aea2a7dcfde9d2684f6b
Reviewed-on: https://chromium-review.googlesource.com/349870
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>