audio: Fix _GetPIDByName function.
We use 'ps name' to get PID, but it is failed on the new kernel. We will
get 'bad pid' result. So we change to use 'pidof name'.
We also use toybox to use the command. So we can use the same logic in
alsa and tinyalsa part.

BUG=None
TEST=Manually tested

Change-Id: Id5791e327b5771d322ff50a250fe73fb097bb45f
Reviewed-on: https://chromium-review.googlesource.com/347564
Commit-Ready: Hsu Wei-Cheng <mojahsu@chromium.org>
Tested-by: Hsu Wei-Cheng <mojahsu@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>