Add option to interrupt process in StartDaemonThread
Sometimes we need to run a function on a separated thread. However, when
the thread crashes with unexpected expection, which is usuall caused by DUT
problem, the main process keeps alive and may hang forever.  Therefore, in
such case we may want to crash the entire process and log execption.

Add an option in StartDaemonThread to interrupt the process when thread
crashes.

BUG=chromium:621380
TEST=Apply to camera test and manually tested.

Change-Id: I6819f4c1262300caeb41bd71ee2282a79c17ac8e
Reviewed-on: https://chromium-review.googlesource.com/353921
Commit-Ready: Shun-Hsing Ou <shunhsingou@chromium.org>
Tested-by: Shun-Hsing Ou <shunhsingou@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>