dkps: fix wrong decryption data when using passphrase
Given a passphrase string that includes newline characters, python-gnupg
reads only the first line (excluding the first newline character), and
prepend the rest to the data. That is, if passphrase is 'abc\ndef', and
the original data is 'xyz'. The data after encryption and decryption
will become '\ndefxyz', not 'xyz'. This is clearly wrong and not
desired, since it may invalidate the format of the original data.

So we should read only the first line (excluding the newline character).
This also complies with the behavior of GnuPG when reading passphrase
from a file.

BUG=None
TEST=Unit tests & manually tested

Change-Id: I6aaa1e2b028cd2b62a7dafc66e675469aea7e812
Reviewed-on: https://chromium-review.googlesource.com/356221
Commit-Ready: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>