init: add function "is_enabled"
In startup, we have two ways to disable a main app:
  1. $ touch main.d/disable-APP
  2. $ chmod -x main.d/APP.sh

To make it easier to determine if a app is enabled or not, the logic is
moved to a single function "is_enabled".
In factory_init, "is_enabled" is used to determine if goofy is enabled
or not.

BUG=chrome-os-partner:30239
TEST=manual

Change-Id: I2e4e78eb52c999dc51a46d31524a2ab03fe5b937
Reviewed-on: https://chromium-review.googlesource.com/357750
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>