installer: add argument "--apps".
If we are using station mode, we might want to install startup scripts
on DUT, but don't want it to start goofy but maybe some DUT initializing
script.

The new argument '--apps' takes a comma separated string which defines
enabling each main apps or not.
For example:

  --apps="-goofy,+whale_servo"

will disable goofy and enable whale_servo after installation.
If an app is not listed, then the enabling status will be left as is.

BUG=chrome-os-partner:30239
TEST=manual

Change-Id: I078de66507b363f109ad1e296edb1fa445640173
Reviewed-on: https://chromium-review.googlesource.com/357751
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>