instalog: add ping-pong functionality to rpc plugins
It is a waste of resources for output_rpc to repeatedly
prepare a batch of events to send, when the host running
the input_rpc plugin may not be available.

Introduce a "Pong" function on input_rpc, which is first
checked by output_rpc before preparing a batch of events
to transmit.

Also, reduce the verbosity of input_rpc, output_rpc, and
output_stdout.

BUG=chromium:609069
TEST=Locally

Change-Id: I7b5a3e39e943aaf97959d2e3ae098af28f33d97c
Reviewed-on: https://chromium-review.googlesource.com/359120
Commit-Ready: Joel Kitching <kitching@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Chun-ta Lin <itspeter@chromium.org>