robot: add function to set logger for the algorithm class
Sometimes it takes pretty long time to compute the result, e.g. camera
calibration. And it's easier to debug to print logs to the console
instead of factory.log. Add an options in the algorithm class of the
robot movement test, so we can print log to console instead of
factory.log

BUG=None
TEST=None

Change-Id: I1cbd796cf9182132783a43c0700118e60c3d3fa3
Reviewed-on: https://chromium-review.googlesource.com/360297
Commit-Ready: Shun-Hsing Ou <shunhsingou@chromium.org>
Tested-by: Shun-Hsing Ou <shunhsingou@chromium.org>
Reviewed-by: Hsu Wei-Cheng <mojahsu@chromium.org>