invocation: move the timing to resolve dargs
In test invocation._run, we log test.dargs before running the test.
However, the logged dargs might not be the actual dargs that is used to
run test, because it need to be resolved by ResolveTestArgs function.

In this CL, we resolve the test args first, log the resolved value,
and call the test eventually.

BUG=chromium:609069
TEST=manual test

Change-Id: I676461944f85db67f07f4fc8cf3d4358fdaefd2a
Reviewed-on: https://chromium-review.googlesource.com/362740
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>