misc: Add standard CrOS testing ssh keys.
Some tests do need to access DUT (especially when doing station-based
testing via SSH Link) with CrOS testing SSH keys. Currently we have a
copy of the keys in setup/ folder but that is not available on station
or DUT. Few projects then try to add keys in board/ folder, which is not
necessary.

This change adds a 'sshkeys' folder into misc/ so every toolkit
installation will have it.

BUG=none
TEST=make test

Change-Id: I248040a3706a2e1100357f6a107079f516c35546
Reviewed-on: https://chromium-review.googlesource.com/364430
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>