Ignore failures in HWTest and VMTest in the commit queue.
Right now, factory changes are rejected if the HWTest or VMTest
stages fail. This is unnecessary as factory changes are very unlikely
to break these stages, because the factory code was not executed in
HWTest and VMTest stages. All factory stuff were only built and archived
in ChromeOS build infrastructure.

BUG=chromium:633833
TEST=make test

Change-Id: I89e0eb43c99a0d20421edf7c88191b287c1cf7b6
Reviewed-on: https://chromium-review.googlesource.com/365437
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>